Refactored & styled the bottom footer
This removes the opacity and switches to flexbox layout. That means that
instead of us having to manually size things and float them left and
right, the browser will just space them out as fits.

The immediate consequence is that all of the text in the center is
visible and readable again.